Improve Panel
