+++ apps/climatemappedafrica/src/components/HURUmap/Panel/MobilePanel/SubcategoryList.js
@@ -1,6 +1,5 @@
-import { Popover, Button } from "@mui/material";
-import makeStyles from "@mui/styles/makeStyles";
-import clsx from "clsx";
+/* eslint-env browser */
+import { Box, Button, Popover } from "@mui/material";
 import PropTypes from "prop-types";
 import React from "react";
 
@@ -8,52 +7,7 @@ import Caret from "@/climatemappedafrica/assets/icons/caret.svg";
 import Link from "@/climatemappedafrica/components/Link";
 import slugify from "@/climatemappedafrica/utils/slugify";
 
-const useStyles = makeStyles(({ palette, typography, zIndex }) => ({
-  root: {
-    position: "sticky",
-    top: 64 + 60, // below navbar and category list tab
-    zIndex: zIndex.appBar,
-  },
-  caretContainer: {
-    alignItems: "center",
-    backgroundColor: palette.grey.light,
-    display: "flex",
-    height: typography.pxToRem(16),
-    justifyContent: "center",
-    width: "100%",
-    "&:active,&:hover, &:focus, &:focus-within": {
-      backgroundColor: palette.grey.light,
-    },
-    "&:after": {
-      display: "none",
-    },
-  },
-  paper: {
-    width: "100%",
-    left: "0 !important",
-    right: 0,
-    maxWidth: "unset",
-    backgroundColor: "#F0F0F0",
-    borderRadius: 0,
-    boxShadow: "unset",
-    paddingTop: typography.pxToRem(9),
-    paddingBottom: typography.pxToRem(24),
-  },
-  title: {
-    width: "100%",
-    textAlign: "center",
-    marginBottom: typography.pxToRem(10),
-    display: "block",
-    cursor: "pointer",
-    color: typography.caption.color,
-  },
-  notSelected: {
-    opacity: 0.2,
-  },
-}));
-
 function SubcategoryList({ items }) {
-  const classes = useStyles();
   const [anchorEl, setAnchorEl] = React.useState(null);
   const [selectedIndex, setSelectedIndex] = React.useState(0);
 
@@ -65,7 +19,6 @@ function SubcategoryList({ items }) {
   };
   const handleSelect = (event) => {
     setSelectedIndex(Number(event.target.dataset.index));
-    /* eslint-env browser */
     document
       .getElementById(slugify(event.target.dataset.title))
       .scrollIntoView({ behavior: "smooth" });
@@ -77,12 +30,33 @@ function SubcategoryList({ items }) {
     return null;
   }
   return (
-    <div className={classes.root}>
-      <Button className={classes.caretContainer} onClick={handleCaretClick}>
+    <Box
+      sx={({ zIndex }) => ({
+        position: "sticky",
+        top: 64 + 60, // below navbar and category list tab
+        zIndex: zIndex.appBar,
+      })}
+    >
+      <Button
+        onClick={handleCaretClick}
+        sx={({ palette, typography }) => ({
+          alignItems: "center",
+          backgroundColor: palette.grey.light,
+          display: "flex",
+          height: typography.pxToRem(16),
+          justifyContent: "center",
+          width: "100%",
+          "&:active,&:hover, &:focus, &:focus-within": {
+            backgroundColor: palette.grey.light,
+          },
+          "&:after": {
+            display: "none",
+          },
+        })}
+      >
         <Caret />
       </Button>
       <Popover
-        classes={{ paper: classes.paper }}
         open={open}
         anchorEl={anchorEl}
         onClose={handleClose}
@@ -94,6 +68,18 @@ function SubcategoryList({ items }) {
           vertical: "top",
           horizontal: "center",
         }}
+        sx={({ typography }) => ({
+          "& .MuiPopover-paper": {
+            backgroundColor: "#F0F0F0",
+            borderRadius: 0,
+            boxShadow: "unset",
+            left: "0",
+            maxWidth: "100%",
+            paddingTop: typography.pxToRem(9),
+            paddingBottom: typography.pxToRem(24),
+            right: 0,
+          },
+        })}
       >
         {items.map(({ title }, index) => (
           <Link
@@ -104,15 +90,26 @@ function SubcategoryList({ items }) {
             href={`#${slugify(title)}`}
             underline="none"
             variant="caption"
-            className={clsx(classes.title, {
-              [classes.notSelected]: selectedIndex !== index,
-            })}
+            sx={[
+              ({ typography }) => ({
+                color: typography.caption.color,
+                cursor: "pointer",
+                display: "block",
+                marginBottom: typography.pxToRem(10),
+                opacity: 0.2,
+                textAlign: "center",
+                width: "100%",
+              }),
+              selectedIndex === index && {
+                opacity: 1.0,
+              },
+            ]}
           >
             {title}
           </Link>
         ))}
       </Popover>
-    </div>
+    </Box>
   );
 }
 SubcategoryList.propTypes = {
