+++ apps/climatemappedafrica/src/components/HURUmap/Panel/Panel.snap.js
@@ -1,0 +1,9 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`<Panel /> renders unchanged 1`] = `
+<div>
+  <div
+    class="MuiBox-root css-0"
+  />
+</div>
+`;
