+++ apps/climatemappedafrica/src/components/HURUmap/Panel/index.js
@@ -1,34 +1,3 @@
-import { Box } from "@mui/material";
-import React from "react";
-
-import DesktopPanel from "./DesktopPanel";
-import MobilePanel from "./MobilePanel";
-
-function Panel(props) {
-  return (
-    <>
-      <Box
-        sx={{
-          display: {
-            xs: "block",
-            lg: "none",
-          },
-        }}
-      >
-        <MobilePanel {...props} />
-      </Box>
-      <Box
-        sx={{
-          display: {
-            xs: "none",
-            lg: "block",
-          },
-        }}
-      >
-        <DesktopPanel {...props} />
-      </Box>
-    </>
-  );
-}
+import Panel from "./Panel";
 
 export default Panel;
